refactor(multiverse): Move the details views under the room view module
