+++ labs/multiverse/src/widgets/mod.rs
@@ -1,7 +1,4 @@
-pub mod events;
 pub mod help;
-pub mod linked_chunk;
-pub mod read_receipts;
 pub mod recovery;
 pub mod room_list;
 pub mod room_view;
